Formatting improvements to csound.py, compliments of PyCharm.
Also, more work on a python-based lilypond parser-- but this will probably be put on hold since I've discovered lilypond's "event-listener" feature. Lilypond.py will be reworked to make use of this, but I'm backing up my previous work as "lp-lexer.py" so it will be easy to switch back to that path if necessary.

Finally, add a lilypond test file.